src/backend/requirements.txt
# Core dependencies for news collection - Optimized versions
aiohttp>=3.9.0,<4.0.0
beautifulsoup4>=4.12.0,<5.0.0
feedparser>=6.0.10,<7.0.0
python-dateutil>=2.8.2,<3.0.0
lxml>=5.0.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0

# Enhanced performance dependencies
aiofiles>=23.0.0,<24.0.0
uvloop>=0.19.0,<1.0.0; sys_platform != "win32"
orjson>=3.9.0,<4.0.0

# Optional ML dependencies (for advanced processing)
# transformers>=4.21.0
# torch>=1.12.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0

# AI and ML dependencies - LangChain with Groq
langchain-groq>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
groq>=0.4.0,<1.0.0

# Swarm Intelligence Framework Dependencies
llama-index-core>=0.12.0
llama-index-workflows>=0.2.0

# Consensus Algorithm Dependencies
scikit-learn>=1.3.0
numpy>=1.24.0

# Enhanced Monitoring and Observability
rich>=13.7.0
structlog>=24.1.0
prometheus-client>=0.20.0
datadog>=0.49.0

# Additional optimization tools
memory-profiler>=0.61.0,<1.0.0
psutil>=5.9.0,<6.0.0